pyproject.toml
[tool.poetry]
name = "eilisch"
version = "0.1.0"
description = ""
authors = ["AtliAsmunds <atliasmunds@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.66.5"
pandas = "^2.2.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
aiohttp = "^3.10.5"
notebook = "^7.2.2"
matplotlib = "^3.9.2"
openai = "^1.52.0"
python-dotenv = "^1.0.1"
numbers-parser = "^4.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
